in order to ease developing multiple modules in eclipse, it's wiser that each bean context file has an unique name.
Despite the main application context loads the childs with
"classpath*:":

<import resource="classpath*:/eu/dnetlib/**/applicationContext.xml"/>

which handles well the situation that a given file is present in
multiple jars, however it has problems when the other modules are
linked as project dependencies in eclipse. The resources will not
presented packaged in jars during runtime and only one
applicationContext.xml file will be arbitrarily selected.

An easy solution is to rename each module's applicationContext using
an unique suffix, for example the module name itself. 

<import resource="classpath*:/eu/dnetlib/**/applicationContext*.xml"/>

This may seem a bit dangerous but the problem only arises if the user
whishes to cut build times by working with eclipse directly.