Modified CQL parser to allow only symbol comparators in relations (allow "f=v", but not "a op b", where op is index-interpreted). Added *very limited* CQL tests.
Modified CQL parser to stop on first error. Used new
runtime exception that forces stop on error (this is
the official antlr way to do this).

Added type field in syntax objects to avoid "instance
of" tests.

Added contain functionality: if query a contains query
b, meaning results from b are a subsets of a results.