CSW-SEARCH-STATUS-TIMESTAMP
Each CSW GetRecords harvesting is featured by a unique time stamp. This is the
same times tamp recorded on each "<csw:GetRecordsResponse>" xml response that is
correlated with the same "GetRecords" CSW request. Consequently, the unique
request id and the search status time stamp are related with an 1-1 relation.

An example element is given below.

<csw:SearchStatus timestamp="2012-05-21T17:53:43Z"/>